+++ src/helpers/utils.helpers.ts
@@ -24,11 +24,15 @@ export function isValidType<T>(key: string, obj: unknown): obj is T {
   return key in obj;
 }
 
-export function parseJson<T>(json: string): T {
+export function parseJson<T>(json?: string | null, throwError = true): T {
   try {
-    const parsed = JSON.parse(json);
+    const parsed = JSON.parse(json ?? '');
     return parsed;
   } catch (error) {
-    throw new Error(`Could not parse JSON: ${json}`);
+    if (throwError) {
+      throw new Error('Could not parse JSON');
+    } else {
+      return {} as T;
+    }
   }
 }
