+++ src/schema/index.ts
@@ -10,7 +10,6 @@ import * as zone from './zone';
 import * as transaction from './transaction';
 import * as topology from './topology';
 import * as workflow from './workflow';
-import * as executedWorkflow from './executed-workflow';
 import * as executedWorkflowTask from './task';
 
 export default makeSchema({
@@ -40,7 +39,6 @@ export default makeSchema({
     transaction,
     topology,
     workflow,
-    executedWorkflow,
     executedWorkflowTask,
   ],
   sourceTypes: {
